% File src/library/stats4/man/AIC-methods.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{AIC-methods}
\docType{methods}
\alias{AIC-methods}
\alias{AIC,ANY-method}
\title{Methods for Function `AIC' in Package `stats4'}
\description{
  Calculate the Akaike information criterion for a fitted model object.
}
\section{Methods}{
  \describe{
    \item{\code{signature(object = "ANY")}}{Generic function: see
      \code{\link[stats]{AIC}}.}
  }
}
\keyword{methods}
