% File src/library/stats4/man/logLik-methods.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{logLik-methods}
\docType{methods}
\alias{logLik-methods}
\alias{logLik,ANY-method}
\alias{logLik,mle-method}
\title{Methods for Function `logLik' in Package `stats4'}
\description{
  Extract the maximized log-likelihood from \code{"mle"} objects.
}
\section{Methods}{\describe{

\item{object = "ANY"}{Generic function: see \code{\link[stats]{logLik}}.}

\item{object = "mle"}{Extract log-likelihood from the fit.}
}}
\keyword{methods}
