% File src/library/stats4/man/profile-methods.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{profile-methods}
\docType{methods}
\alias{profile-methods}
\alias{profile,ANY-method}
\alias{profile,mle-method}
\title{Methods for Function profile in Package `stats4'}
\description{Profile likelihood}
\section{Methods}{\describe{

\item{fitted = "ANY"}{Generic function}

\item{fitted = "mle"}{Profile the likelihood in the vicinity of the optimum}
}}
\keyword{methods}
