\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,mle-method}
\alias{show,summary.mle-method}
\title{Methods for Function show in Package `stats4'}
\description{Show objects of classes \code{mle} and \code{summary.mle}}
\section{Methods}{\describe{

\item{object = "mle"}{Print simple summary of \code{mle} object. Just
  the coefficients and the call.}

\item{object = "summary.mle"}{Shows call, table of coefficients and
  standard errors, and -2 log L}
}}
\keyword{methods}
