\name{stats4.package}
\alias{stats4.package}
\alias{stats4}
\docType{package}
\title{
Statistical Functions using S4 Classes
}
\description{
Statistical Functions using S4 classes.
}
\details{
This package contains functions and classes for statistics
using the \link[methods:methods.package]{S version 4} class system.

The methods currently support \link[stats4:mle]{maximum likelihood}, 
including \link[stats4:AIC-methods]{AIC} and \link{BIC} for model selection.

}
\author{
R Development Core Team and contributors worldwide

Maintainer: R Core Team <R-core@r-project.org>
}
\keyword{ package }
\keyword{ models }