\name{vcov-methods}
\docType{methods}
\alias{vcov-methods}
\alias{vcov,ANY-method}
\alias{vcov,mle-method}
\title{Methods for Function `vcov' in Package `stats4'}
\description{
  Extract the approximate variance-covariance matrix from \code{"mle"} objects.
}
\section{Methods}{\describe{

\item{object = "ANY"}{Generic function: see \code{\link[stats]{vcov}}.}

\item{object = "mle"}{Extract the estimated variance-covariance matrix
  for the estimated parameters (if any).}
}}
\keyword{methods}
