\name{AIC-methods}
\docType{methods}
\alias{AIC-methods}
\alias{AIC,ANY-method}
\title{Methods for Function `AIC' in Package `stats4'}
\description{
  Calculate the Akaike information criterion for a fitted model object.
}
\section{Methods}{
  \describe{
    \item{object = "ANY"}{Generic function: see \code{\link[stats]{AIC}}.}
  }
}
\keyword{methods}
