\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,mle-method}
\title{Methods for Function summary in Package `stats4'}
\description{Summarize objects}
\section{Methods}{\describe{

\item{object = "ANY"}{Generic function}

\item{object = "mle"}{Generate a summary as an object of class
  "summary.mle", containing estimates, asymptotic SE, and value of
  -2 log L}
}}
\keyword{methods}
